Update dependencies and fix uint/uint64 type mismatches from gorm v1.31

+++ pkg/types/auth.types.go
@@ -40,7 +40,7 @@ func UserFromDal(user *dal.User) *User {
 	achievements := Achievements(activities)
 
 	return &User{
-		ID:           user.ID,
+		ID:           uint64(user.ID),
 		Name:         user.Name,
 		Email:        user.Email,
 		Activities:   activities,
